S-01041 PID controller - basic PID implementation: 1. moved some static definitions into Config (to be continued); 2. implemented basic proportional heading controller; 3. cleaned up existing altitude/pitch controller (has to be revised); 4. added boundaries setter to PID implementation.

+++ src/data/DesiredData.h
@@ -30,7 +30,7 @@ class DesiredData
 		int depth = 0;
 
 		// desired direction (degrees relative to North)
-		float heading = 0;
+		int heading = 0;
 
 		/* desired horizontal (forward/reverse) speed (predefined levels, from 0 to 8) */
 		int speed = 0;
